Update patch set 2
Patch Set 2:

Yeah, so http://www.linux-magazine.com/Issues/2013/156/exFAT-Filesystem says both exFAT fuse and nofuse have the risk of legal concerns. IMO, we should drop exfat fuse entirely, and make the decision on whether not to include it for devices that are licensed only, or for all devices. I lean towards licensed devices only, but that's not my call. Shipping exfat fuse is a disservice to our users, it's tremendously broken (new UUID on every boot) and horrible to use (we just switched to sdcardfs to get rid of FUSE for performance, and exfat forces us to go back), in my opinion.

Patch-set: 2