Update patch set 3
Patch Set 3:

Right. This is slightly better than the existing situation of shipping exfat fuse for all, but I guess we're still not 100% clear on whether or not the device shipping it on stock will cover us using it on Lineage. I suppose the alternative is to kill exfat support for everyone, but I'm -1 to that, that's crappy for our users.

Patch-set: 3