Update patch set 4
Patch Set 4:

I don't really like the idea of "make an exception for new devices to add rules for old devices". New devices/methods should be the reference, the ones which rules are supposed to be applied to. Especially given that prebuilt kernel is the method used and recommended by AOSP now.

Patch-set: 4
CC: Jesse Chan <14934@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>