Update patch set 4
Patch Set 4: Code-Review+2

> I don't really like the idea of "make an exception for new devices
 > to add rules for old devices". New devices/methods should be the
 > reference, the ones which rules are supposed to be applied to.
 > Especially given that prebuilt kernel is the method used and
 > recommended by AOSP now.

AOSP always used/recommended prebuilt kernel. We always built it from source. This is the rule that is currently enforced on all the devices, so we are documenting it. We can discuss adding exceptions when such devices are ready to ship and comply with all the rest of the charter

Patch-set: 4
Reviewer: Michael Bestas <4550@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+2