Update patch set 4
Patch Set 4:

> AOSP always used/recommended prebuilt kernel. We always built it
 > from source. This is the rule that is currently enforced on all the
 > devices, so we are documenting it. We can discuss adding exceptions
 > when such devices are ready to ship and comply with all the rest of
 > the charter

Why not just add an exception alongside this commit for Treble-shipped devices?

Or a note that they have grounds to ask for an exception?

We should definitely note it publicly IMO.

> My intention is to add a section on treble, because I assume a few of these rules may be irrelevant (such as this)

This. Please add me when you do, I'm definitely interested to see that.

> Then those maintainers should learn the skill, or work on devices with with kernel source available. I mean we don’t dumb down requirements (aka selinux is hard, but we still require it) for any other reason, so why allow this?

I am with you that RE is a skill to learn, though honestly, it is not a skill I think we can ask of the typical maintainer.

Patch-set: 4
CC: Nolen Johnson <16185@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>