Update patch set 4
Patch Set 4:

"I am with you that RE is a skill to learn, though honestly, it is not a skill I think we can ask of the typical maintainer." 

Most maintainers have source code, I think we should be deliberately selective when we force this, as it will just make sure only the brightest/most dedicated maintainers add their devices to hudson. I think it's very reasonable to require maintainers to do this if they CHOOSE to work on a device without source.

Patch-set: 4